# contracts/openapi/platform-v1.yaml
openapi: 3.1.0
info:
  title: Wealth Sandbox Platform API
  version: 1.0.0
  description: Canonical contract consumed by the mobile application.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Current platform API origin
paths:
...
  /api/v1/mydata/connections:
    post:
      operationId: createMyDataConnection
      summary: Create the single synthetic institution connection
      description: Requires `financial.write`; the external customer identifier is never returned.
      security:
        - BearerAuth: []
      tags: [MyData]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateConnectionRequest'
      responses:
        '201':
          description: Synthetic institution connection created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Connection'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: An active connection already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    get:
      operationId: listMyDataConnections
      summary: List the current user's institution connections
      description: Requires `financial.read`.
      security:
        - BearerAuth: []
      tags: [MyData]
      responses:
        '200':
          description: Zero or one connection in MVP.
          content:
            application/json:
              schema:
                type: array
                maxItems: 1
                items:
                  $ref: '#/components/schemas/Connection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/mydata/syncs:
    post:
      operationId: createMyDataSync
      summary: Queue a manual synthetic institution synchronization
      description: Requires `financial.write`. An already-active sync is returned with HTTP 200.
      security:
        - BearerAuth: []
      tags: [MyData]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSyncRequest'
      responses:
        '200':
          description: Existing active sync.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sync'
        '202':
          description: New sync queued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sync'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/mydata/syncs/{syncId}:
    get:
      operationId: getMyDataSync
      summary: Read a user-owned sync status
      description: Requires `financial.read` and ownership.
      security:
        - BearerAuth: []
      tags: [MyData]
      parameters:
        - name: syncId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Current sync status and normalized counts.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sync'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/simulations:
    post:
      operationId: createSimulation
      summary: Run and persist a deterministic synthetic financial simulation
      description: Requires `simulation.execute`. The server generates and stores the seed.
      security:
        - BearerAuth: []
      tags: [Simulation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSimulationRequest'
      responses:
        '201':
          description: Completed and persisted synthetic simulation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Simulation'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/simulations/{simulationId}:
    get:
      operationId: getSimulation
      summary: Read one persisted user-owned simulation
      description: Requires `simulation.execute`; non-owned resources return 404.
      security:
        - BearerAuth: []
      tags: [Simulation]
      parameters:
        - name: simulationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Persisted simulation result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Simulation'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/dev/scenario:
    put:
      operationId: setDeveloperScenario
      summary: Set the local/demo simulator scenario
      description: Local is relaxed; demo requires `scenario.admin`. The module is absent in production.
      security:
        - {}
        - BearerAuth: []
      tags: [Developer]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetDeveloperScenarioRequest'
      responses:
        '200':
          description: Applied global simulator scenario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeveloperScenarioResponse'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/dev/dataset/reset:
    post:
      operationId: resetDeveloperDataset
      summary: Reset the local/demo synthetic simulator dataset
      description: Local is relaxed; demo requires `scenario.admin`. The module is absent in production.
      security:
        - {}
        - BearerAuth: []
      tags: [Developer]
      responses:
        '200':
          description: Deterministic synthetic baseline restored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeveloperResetResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: Bearer token is missing, expired, or cannot be verified.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Bearer token does not include the required scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    ValidationFailed:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource does not exist or is not owned by the caller.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  schemas:
    CreateConnectionRequest:
      type: object
      additionalProperties: false
      required: [institutionCode, consentExpiresAt]
      properties:
        institutionCode:
          type: string
          const: SYNTH_WEALTH_001
        consentExpiresAt:
          type: string
          format: date-time
    Connection:
      type: object
      additionalProperties: false
      required:
        - connectionId
        - institutionCode
        - status
        - consentExpiresAt
        - lastSuccessfulSyncAt
      properties:
        connectionId:
          type: string
          format: uuid
        institutionCode:
          type: string
          const: SYNTH_WEALTH_001
        status:
          type: string
          enum: [ACTIVE, REVOKED, EXPIRED]
        consentExpiresAt:
          type: string
          format: date-time
        lastSuccessfulSyncAt:
          type: [string, 'null']
          format: date-time
    CreateSyncRequest:
      type: object
      additionalProperties: false
      required: [connectionId]
      properties:
        connectionId:
          type: string
          format: uuid
    Sync:
      type: object
      additionalProperties: false
      required:
        - syncId
        - connectionId
        - status
        - createdAt
        - startedAt
        - completedAt
        - counts
        - errorCode
      properties:
        syncId:
          type: string
          format: uuid
        connectionId:
          type: string
          format: uuid
        status:
          type: string
          enum: [QUEUED, FETCHING, RAW_STORED, NORMALIZING, COMPLETED, FAILED]
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: [string, 'null']
          format: date-time
        completedAt:
          type: [string, 'null']
          format: date-time
        counts:
          $ref: '#/components/schemas/SyncCounts'
        errorCode:
          type: [string, 'null']
    SyncCounts:
      type: object
      additionalProperties: false
      required: [rawRecords, accounts, holdings, transactions]
      properties:
        rawRecords:
          type: integer
          minimum: 0
        accounts:
          type: integer
          minimum: 0
        holdings:
          type: integer
          minimum: 0
        transactions:
          type: integer
          minimum: 0
    Money:
      type: string
      pattern: '^-?[0-9]+\.[0-9]{4}$'
...
    CreateSimulationRequest:
      type: object
      additionalProperties: false
      required:
        - initialAssets
        - monthlyContribution
        - durationMonths
        - targetAmount
        - allocation
      properties:
        initialAssets:
          $ref: '#/components/schemas/NonNegativeMoneyInput'
        monthlyContribution:
          $ref: '#/components/schemas/NonNegativeMoneyInput'
        durationMonths:
          type: integer
          minimum: 1
          maximum: 600
        targetAmount:
          $ref: '#/components/schemas/NonNegativeMoneyInput'
        allocation:
          type: array
          minItems: 1
          maxItems: 3
          items:
            type: object
            additionalProperties: false
            required: [assetClass, weight]
            properties:
              assetClass:
                type: string
                enum: [CASH, BOND, EQUITY]
              weight:
                type: number
                minimum: 0
                maximum: 1
    NonNegativeMoneyInput:
      type: string
      pattern: '^[0-9]+(?:\.[0-9]{1,4})?$'
    Simulation:
      type: object
      additionalProperties: false
      required:
        - simulationId
        - engineVersion
        - assumptionSetVersion
        - currency
        - goalProbability
        - finalValue
        - series
        - disclaimer
      properties:
        simulationId:
          type: string
          format: uuid
        engineVersion:
          type: string
          const: 1.0.0
        assumptionSetVersion:
          type: string
          const: SYNTHETIC_V1
        currency:
          type: string
          const: KRW
        goalProbability:
          type: number
          minimum: 0
          maximum: 1
        finalValue:
          $ref: '#/components/schemas/SimulationPercentiles'
        series:
          type: array
          minItems: 2
          maxItems: 601
          items:
            $ref: '#/components/schemas/SimulationPoint'
        disclaimer:
          type: string
          const: Synthetic financial simulation for technical demonstration only.
    SimulationPercentiles:
      type: object
      additionalProperties: false
      required: [p10, p50, p90]
      properties:
        p10:
          $ref: '#/components/schemas/Money'
        p50:
          $ref: '#/components/schemas/Money'
        p90:
          $ref: '#/components/schemas/Money'
    SimulationPoint:
      type: object
      additionalProperties: false
      required: [month, p10, p50, p90]
      properties:
        month:
          type: integer
          minimum: 0
          maximum: 600
        p10:
          $ref: '#/components/schemas/Money'
        p50:
          $ref: '#/components/schemas/Money'
        p90:
          $ref: '#/components/schemas/Money'
...
    DeveloperScenarioMode:
      type: string
      enum:
        [
          NORMAL,
          TIMEOUT,
          HTTP_500,
          MALFORMED_RESPONSE,
          ORDER_REJECT,
          ORDER_UNKNOWN_THEN_FILLED,
        ]
    SetDeveloperScenarioRequest:
      type: object
      additionalProperties: false
      required: [mode]
      properties:
        mode:
          $ref: '#/components/schemas/DeveloperScenarioMode'
        correlationScope:
          type: string
          const: CURRENT_USER
    DeveloperScenarioResponse:
      type: object
      additionalProperties: false
      required: [mode, scope]
      properties:
        mode:
          $ref: '#/components/schemas/DeveloperScenarioMode'
        scope:
          type: string
          const: GLOBAL
    DeveloperResetResponse:
      type: object
      additionalProperties: false
      required: [datasetVersion, scenarioMode, syntheticData]
      properties:
        datasetVersion: { type: string, minLength: 1 }
        scenarioMode: { type: string, const: NORMAL }
        syntheticData: { type: boolean, const: true }
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required:
        - type
        - title
        - status
        - code
        - detail
        - traceId
        - retryable
        - fieldErrors
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
        detail:
          type: string
        traceId:
          type: string
        retryable:
          type: boolean
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - field
              - code
              - message
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
...